refactor(napi/parser): move raw transfer code into separate directory (#11583)
Pure refactor. Move code for raw transfer into separate directory. This will enable exposing some internals to benchmarks without exposing them publicly from the package. Also lazy deserialization is going to add a lot more code - don't want `index.js` to get too bloated.

Also make main entry point `index.js` run in strict mode (doesn't much matter, but it's good practice).
